refactor: update dataset listing parameters

+++ scrapeless/services/storage/dataset.go
@@ -28,8 +28,8 @@ func (s *Dataset) ListDatasets(ctx context.Context, page int64, pageSize int64, 
 	datasets, err := storage.ClientInterface.ListDatasets(ctx, &models.ListDatasetsRequest{
 		ActorId:  &env.GetActorEnv().ActorId,
 		RunId:    &env.GetActorEnv().RunId,
-		Page:     int(page),
-		PageSize: int(pageSize),
+		Page:     page,
+		PageSize: pageSize,
 		Desc:     desc,
 	})
 	if err != nil {
@@ -51,7 +51,7 @@ func (s *Dataset) ListDatasets(ctx context.Context, page int64, pageSize int64, 
 	}
 	return &ListDatasetsResponse{
 		Items: itemArray,
-		Total: int64(datasets.Total),
+		Total: datasets.Total,
 	}, nil
 }
